vayu: power-libperfmgr: reduce default GPU idle timer to 70 ms
We have commits in kernel to set default GPU idle timeout to 64ms
so that it aligns with panel ULPS timeout value. However, the
actual idle timeout value at runtime will be 70ms because the
kernel uses 100HZ jiffies. Change the power hint so that idle
timeout returns to the default 70ms after a LAUNCH boost.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Signed-off-by: DennySPb <dennyspb@gmail.com>
Change-Id: I655fa28f6b721d659b3a0d0941c7d8674812ddf3
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Rafiester <rafi.haqqi@gmail.com>